Set the manually the env vars from env_var_name parameters
The parameters of type env_var_name are holding the name of
environment variable, not its value.
Most probably due to the specific handling of parameters of
type env_var_name, there is a complication in passing theif
values to a script. Their values are not passed to a script
environment as are passed parameters of other types.
The usual syntax is not working.
So we are setting them manually in the environment.